+++ firebase_core/_firebase_database_collection_vm/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2+1
+
+- constrain test, collection, meta dependency
+
 ## 0.0.2
 
 - Update README