+++ firebase_core/firebase_core_dart/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Implement new platform interface contract
+
 ## 0.0.1
 
 * Initial release.